Add a slightly broken implementation of client side movement smoothing with server position verification

+++ core/src/com/test/kt/SimulationSnapshot.java
@@ -4,7 +4,7 @@ package com.test.kt;
  * Created by chrisb on 10/05/2017.
  */
 public class SimulationSnapshot {
-    public float timestamp;
+    public long timestamp;
     public float px = 0;
     public float py = 0;
 }
